# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel OpenApi",
        "description": "L5 Swagger OpenApi description",
        "contact": {
            "email": "admin@admin.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api/v1",
            "description": "Order Me API"
        }
    ],
    "paths": {
...
        "/admin/room": {
            "get": {
                "tags": [
                    "Room"
                ],
                "summary": "Get list of room",
                "description": "Returns list of room",
                "operationId": "getRooms",
                "parameters": [
                    {
                        "name": "keyword",
                        "in": "query",
                        "description": "user name",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/RoomResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/room/message-not-seen-by-user/{phone}": {
            "get": {
                "tags": [
                    "Room"
                ],
                "summary": "Get room information",
                "description": "Returns room data",
                "operationId": "getMessageNotSeenByUser",
                "parameters": [
                    {
                        "name": "phone",
                        "in": "path",
                        "description": "User Phone",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/RoomResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/shipper/order": {
            "get": {
                "tags": [
                    "Shipper Order"
                ],
                "summary": "Get list of order",
                "description": "Returns list of order",
                "operationId": "getOrdersByShipper",
                "parameters": [
                    {
                        "name": "keyword",
                        "in": "query",
                        "description": "Can search by order code, user name",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "limit size ",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "page size ",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "description": "status of order",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "orderBy",
                        "in": "query",
                        "description": "can sort by multiple field vd :-id,+id,+name,-name,-code, +code",
                        "required": false,
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/OrderResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/shipper/order/{id}": {
            "get": {
                "tags": [
                    "Shipper Order"
                ],
                "summary": "Get order information",
                "description": "Returns order data",
                "operationId": "getOrderShipperById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "order id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/OrderDetailResponse"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Shipper Order"
                ],
                "summary": "Update existing Order",
                "description": "Returns updated location data",
                "operationId": "updateOrderShipper",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Order id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/OrderUpdate"
                            }
                        }
                    }
                },
                "responses": {
                    "202": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/OrderResponse"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "ChatResponse": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "sender": {
                        "$ref": "#/components/schemas/UserResponse"
                    },
                    "content": {
                        "type": "string"
                    },
                    "created_at": {
                        "type": "string"
                    },
                    "updated_at": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
            "LocationResponse": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "address": {
                        "type": "string"
                    },
                    "distance": {
                        "type": "number"
                    },
                    "created_at": {
                        "type": "string"
                    },
                    "updated_at": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
            "LogResponse": {
                "title": "Log Data",
                "description": "Log model",
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "status": {
                        "type": "string"
                    },
                    "change_by": {
                        "$ref": "#/components/schemas/UserResponse"
                    },
                    "created_at": {
                        "type": "string"
                    },
                    "updated_at": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
            "OrderDetailResponse": {
                "title": "Order Detail Response",
                "description": "Order model",
                "properties": {
                    "dishes": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/OrderDishItem"
                        }
                    },
                    "logs": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/LogResponse"
                        }
                    },
                    "total": {
                        "type": "string"
                    },
                    "created_at": {
                        "type": "string"
                    },
                    "updated_at": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "OrderDishItem": {
                "title": "Order Dish Item",
                "description": "Order model",
                "properties": {
                    "id": {
                        "type": "number"
                    },
                    "image": {
                        "type": "string"
                    },
                    "name": {
                        "type": "string"
                    },
                    "slug": {
                        "type": "string"
                    },
                    "quantity": {
                        "type": "number"
                    },
                    "price": {
                        "type": "number"
                    }
                },
                "type": "object"
            },
            "OrderResponse": {
                "title": "Order Request",
                "description": "Order model",
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "status": {
                        "type": "integer"
                    },
                    "code": {
                        "type": "string"
                    },
                    "phone": {
                        "type": "string"
                    },
                    "total": {
                        "type": "string"
                    },
                    "note": {
                        "type": "string"
                    },
                    "location": {
                        "$ref": "#/components/schemas/LocationResponse"
                    },
                    "user": {
                        "$ref": "#/components/schemas/UserResponse"
                    },
                    "vnp_url": {
                        "type": "string"
                    },
                    "location_detail": {
                        "type": "string"
                    },
                    "price_sale": {
                        "type": "integer"
                    },
                    "price_none_sale": {
                        "type": "integer"
                    },
                    "coupon_id": {
                        "type": "integer"
                    },
                    "created_at": {
                        "type": "string"
                    },
                    "updated_at": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "OrderUpdate": {
                "title": "Order Request",
                "description": "Order model",
                "properties": {
                    "status": {
                        "type": "integer"
                    },
                    "phone": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
            "RoomResponse": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "user_phone": {
                        "type": "string"
                    },
                    "user_name": {
                        "type": "string"
                    },
                    "messages": {
                        "$ref": "#/components/schemas/ChatResponse"
                    }
                },
                "type": "object"
            },
...
            "UserResponse": {
                "properties": {
                    "name": {
                        "type": "string"
                    },
                    "phone": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    },
                    "avatar": {
                        "type": "string"
                    },
                    "status": {
                        "type": "integer"
                    },
                    "created_at": {
                        "type": "string"
                    },
                    "updated_at": {
                        "type": "string"
                    },
                    "id": {
                        "type": "integer"
                    }
                },
                "type": "object"
            }
        }
    },
...
}